requirements.txt
mnemonic
python-dotenv